split shaky tests into separate files
