+++ extension/test/server/shaky/test_catchup_exercises.py
@@ -1,0 +1,68 @@
+"""
+This file is part of Arakoon, a distributed key-value store. Copyright
+(C) 2010 Incubaid BVBA
+
+Licensees holding a valid Incubaid license may use this file in
+accordance with Incubaid's Arakoon commercial license agreement. For
+more information on how to enter into this agreement, please contact
+Incubaid (contact details can be found on www.arakoon.org/licensing).
+
+Alternatively, this file may be redistributed and/or modified under
+the terms of the GNU Affero General Public License version 3, as
+published by the Free Software Foundation. Under this license, this
+file is distributed in the hope that it will be useful, but WITHOUT
+ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+nFITNESS FOR A PARTICULAR PURPOSE.
+
+See the GNU Affero General Public License for more details.
+You should have received a copy of the
+GNU Affero General Public License along with this program (file "COPYING").
+If not, see <http://www.gnu.org/licenses/>.
+"""
+
+from .. import system_tests_common as Common
+from arakoon.ArakoonExceptions import *
+import arakoon
+import logging
+import time
+import subprocess
+from threading import Thread
+from nose.tools import *
+
+
+@Common.with_custom_setup(Common.setup_2_nodes_forced_master, Common.basic_teardown)
+def test_catchup_exercises():
+    time.sleep(Common.lease_duration) # ??
+
+    def do_one(n, max_wait):
+        logging.info("do_one(%i,%f)", n, max_wait)
+        Common.iterate_n_times(n, Common.simple_set)
+        logging.info("sets done, stopping")
+        Common.stop_all()
+        logging.info("stopped all nodes")
+        nn = Common.node_names[1]
+        Common.wipe(nn)
+
+        Common.start_all()
+        logging.info("started all")
+
+        counter = 0
+        up2date = False
+        cli = Common.get_client ()
+        while not up2date and counter < max_wait :
+            time.sleep( 1.0 )
+            counter += 1
+            up2date = cli.expectProgressPossible()
+            logging.info("up2date=%s", up2date)
+
+        cli.dropConnections()
+        if counter >= max_wait :
+            raise Exception ("Node did not catchup in a timely fashion")
+
+    n = 20000
+    w = 80 # 250/s should be more than enough even for dss driven vmachines
+    for i in range(5):
+        do_one(n,w)
+        n = n * 2
+        w = w * 2
+
